elearning_project/requirements.txt
asgiref==3.11.1
channels==4.3.2
Django==6.0.2
djangorestframework==3.16.1
pillow==12.1.0
redis==7.1.1
sqlparse==0.5.5
tzdata==2025.3
